Restructure the task related class into a package
