+++ src/main/java/lolok/Storage.java
@@ -1,5 +1,7 @@
 package lolok;
 
+import lolok.task.Task;
+
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
